refactor: move debug-view out of gallery/
The debug overlay renders card chrome — a Lit template in the status bar row,
mounted by card.ts. It sat under gallery/ because the numbers it shows come
from the gallery fetch protocol, but it was the only template in a directory
that is otherwise fetching, caching and backoff.

The counters stay where they are produced (gallery/debug-stats.ts, five
importers there); only the rendering half moves up beside card-template.ts.
